// src/types/member.ts
export interface MemberType {
  id: number;
  name: string;
  phone: string;
  email: string;
  cardNo: string;
  dueDate: string;
  membership: string;
  avatar: string;
  gender?: string;
  dob?: string;
  idType?: string;
  cnic?: string;
  bloodGroup?: string;
  height?: string;
  weight?: string;
  emergencyName?: string;
  emergencyMobile?: string;
  street?: string;
  city?: string;
  country?: string;
  comment?: string;
  joiningDate?: string;
}